extract logging function into its own class
