+++ src/logout/logout.ts
@@ -2,15 +2,13 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 import * as core from "@actions/core";
-import * as exec from "@actions/exec";
 import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
+import { execCommand, getLogMode, logInfo, withLogGroup } from "../logging";
 
 const runnerWindows = "Windows";
 const runnerMacOS = "macOS";
-
-type LogMode = "grouped" | "normal" | "quiet";
 
 async function logout(): Promise<void> {
   try {
@@ -21,23 +19,24 @@ async function logout(): Promise<void> {
       if (runnerOS === runnerMacOS) {
         // The below is required to allow GitHub's post job cleanup to complete.
         logInfo(logMode, "Resetting DNS settings on macOS");
-        await execCommand(logMode, "networksetup", [
-          "-setdnsservers",
-          "Ethernet",
-          "Empty",
-        ]);
-        await execCommand(logMode, "networksetup", [
-          "-setsearchdomains",
-          "Ethernet",
-          "Empty",
-        ]);
+        await execCommand(
+          "networksetup",
+          ["-setdnsservers", "Ethernet", "Empty"],
+          { logMode },
+        );
+        await execCommand(
+          "networksetup",
+          ["-setsearchdomains", "Ethernet", "Empty"],
+          { logMode },
+        );
       }
 
       logInfo(logMode, "🔄 Logging out of Tailscale...");
 
       // Check if tailscale is available first
       try {
-        await execCommand(logMode, "tailscale", ["--version"], {
+        await execCommand("tailscale", ["--version"], {
+          logMode,
           silent: true,
         });
 
@@ -53,7 +52,7 @@ async function logout(): Promise<void> {
         logInfo(logMode, `Running: ${execArgs.join(" ")}`);
 
         try {
-          await execCommand(logMode, execArgs[0], execArgs.slice(1));
+          await execCommand(execArgs[0], execArgs.slice(1), { logMode });
           logInfo(logMode, "✅ Successfully logged out of Tailscale");
         } catch (error) {
           // Don't fail the action if logout fails - it's just cleanup
@@ -74,12 +73,10 @@ async function logout(): Promise<void> {
       logInfo(logMode, "Stopping tailscale");
       try {
         if (runnerOS === runnerWindows) {
-          await execCommand(logMode, "net", ["stop", "Tailscale"]);
-          await execCommand(logMode, "taskkill", [
-            "/F",
-            "/IM",
-            "tailscale-ipn.exe",
-          ]);
+          await execCommand("net", ["stop", "Tailscale"], { logMode });
+          await execCommand("taskkill", ["/F", "/IM", "tailscale-ipn.exe"], {
+            logMode,
+          });
         } else {
           const xdgRuntimeDir =
             process.env.XDG_RUNTIME_DIR ||
@@ -92,9 +89,9 @@ async function logout(): Promise<void> {
             throw new Error("pid file empty");
           }
           // The pid is actually the pid of the `sudo` parent of tailscaled, so use pkill -P to kill children of that parent
-          await execCommand(logMode, "sudo", ["pkill", "-P", pid]);
+          await execCommand("sudo", ["pkill", "-P", pid], { logMode });
           // Clean up DNS and routes.
-          await execCommand(logMode, "sudo", ["tailscaled", "--cleanup"]);
+          await execCommand("sudo", ["tailscaled", "--cleanup"], { logMode });
         }
         logInfo(logMode, "✅ Stopped tailscale");
       } catch (error) {
@@ -105,60 +102,6 @@ async function logout(): Promise<void> {
     // Don't fail the action for post-cleanup issues
     core.warning(`Post-action cleanup error: ${error}`);
   }
-}
-
-function getLogMode(): LogMode {
-  const logMode = core.getInput("log-mode") || "grouped";
-  if (logMode !== "grouped" && logMode !== "normal" && logMode !== "quiet") {
-    throw new Error(
-      `Invalid log-mode "${logMode}". Expected "grouped", "normal", or "quiet".`,
-    );
-  }
-  return logMode;
-}
-
-function logInfo(logMode: LogMode, message: string): void {
-  if (logMode !== "quiet") {
-    core.info(message);
-  }
-}
-
-async function withLogGroup<T>(
-  logMode: LogMode,
-  name: string,
-  fn: () => Promise<T>,
-): Promise<T> {
-  if (logMode !== "grouped") {
-    return fn();
-  }
-
-  core.startGroup(name);
-  try {
-    return await fn();
-  } finally {
-    core.endGroup();
-  }
-}
-
-async function execCommand(
-  logMode: LogMode,
-  commandLine: string,
-  args?: string[],
-  options?: exec.ExecOptions,
-): Promise<number> {
-  const silent = options?.silent || logMode === "quiet";
-  const out = await exec.getExecOutput(commandLine, args, {
-    ...options,
-    silent,
-    ignoreReturnCode: true,
-  });
-  if (out.exitCode !== 0) {
-    if (silent) {
-      process.stderr.write(out.stderr);
-    }
-    throw new Error(`${commandLine} failed with exit code ${out.exitCode}`);
-  }
-  return out.exitCode;
 }
 
 // Run the logout function
